requirements.txt
apipkg==1.5
attrs==20.2.0
click==7.1.2
coverage==5.3
execnet==1.7.1
importlib-metadata==2.0.0
iniconfig==1.0.1
memory-profiler==0.57.0
more-itertools==8.5.0
mysql-connector-python==8.0.21
mysqlclient==2.0.1
packaging==20.4
pluggy==0.13.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
py==1.9.0
Pygments==2.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytest-forked==1.3.0
pytest-xdist==2.1.0
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
six==1.15.0
tabulate==0.8.7
toml==0.10.1
wcwidth==0.2.5
zipp==3.2.0
